refactor: share summary and embedding retry policy
